Closes #2402: Fix metrics page when postgres14 is in use (#2406)

+++ dataverse-webapp/src/main/java/edu/harvard/iq/dataverse/metrics/MetricsServiceBean.java
@@ -211,8 +211,8 @@ public class MetricsServiceBean implements Serializable {
     private List<ChartMetrics> mapToChartMetrics(List<Object[]> result) {
         return result.stream()
                 .map(dm -> new ChartMetrics(
-                        dm[0] instanceof BigDecimal ?  ((BigDecimal) dm[0]).intValue() : ((Integer)dm[0]),
-                        dm[1] instanceof BigDecimal ?  ((BigDecimal) dm[1]).intValue() : ((Integer)dm[1]),
+                        dm[0] instanceof BigDecimal ?  ((BigDecimal) dm[0]).intValue() : ((Double)dm[0]).intValue(),
+                        dm[1] instanceof BigDecimal ?  ((BigDecimal) dm[1]).intValue() : ((Double)dm[1]).intValue(),
                         dm[2] instanceof BigDecimal ?  ((BigDecimal) dm[2]).longValue() : ((Long)dm[2]))
                 )
                 .collect(Collectors.toList());
